requirements.txt
langchain
langchain-openai
streamlit
langchain-community
langchain-core
langchain-chroma
chroma
fastapi[standard]
python-dotenv
pypdf
requests

